Skip initial whitespace when decoding a dict.
This fixes cases where there is whitespace before the '{' that
begins the top-level object. Also, omit the newline before '{' when
writing the top-level object.

Either fix would be sufficient on its own to fix the symptom, which
is that applying multiple merges to the same sjson file didn't work
correctly with the previous change. With this fix, the round-trip
behavior is restored, so multiple edits to the same sjson file work
as expected.