Correctly handle SGG sjson block quotes that end with a quotation mark.
Several of the .sjson files that come with Hades include quotation marks
within triple quotes, ie.

   """"Something Persephone said to Zagreus""""

In these cases, we must consider the last three quotes of a contiguous sequence
of quotes to end the block; if we end with the first three we will almost
certainly get a parsing error.

Mostly affects non-english files in Game/Text .