Better support for mac via command line argument.
On Mac, the game files are in a 'Resources' directory rather than
one named after the game. This makes automatically determining which
game is being modded tricky.

Add a command line flag that can be used to specify the game, so that
the automatic determination can be overridden by the user, this allows
mac users to simply "python modimporter.py --game Hades" instead of having
to rename or modify any files.